+++ src/Client.php
@@ -2,13 +2,17 @@
 
 namespace Kuusamo\Api;
 
+use Kuusamo\Api\Exception\ResponseException;
 use Buzz\Client\FileGetContents;
 use Nyholm\Psr7\Factory\Psr17Factory;
+use Psr\Http\Client\ClientInterface;
 
 class Client
 {
     private $token;
     private $baseUri;
+    private $psr17Factory;
+    private $networkClient;
 
     /**
      * Constructor.
@@ -23,6 +27,28 @@ class Client
     }
 
     /**
+     * Set the PSR-17 factory to use.
+     *
+     * @param Psr17Factory $factory Factory.
+     * @return void
+     */
+    public function setPsr17Factory(Psr17Factory $factory)
+    {
+        $this->psr17Factory = $factory;
+    }
+
+    /**
+     * Set the network client to use.
+     *
+     * @param ClientInterface $client Client.
+     * @return void
+     */
+    public function setNetworkClient(ClientInterface $client)
+    {
+        $this->networkClient = $client;
+    }
+
+    /**
      * Send a request.
      *
      * @param string $method GET, POST, PUT, DELETE
@@ -34,7 +60,7 @@ class Client
     {
         $url = $this->baseUri . $path;
 
-        $factory = new Psr17Factory;
+        $factory = $this->psr17Factory ?: new Psr17Factory;
         $request = $factory->createRequest($method, $url);
         $request = $request->withHeader('Content-Type', 'application/json');
         $request = $request->withHeader('Kuusamo-Key', $this->token->getKey());
@@ -49,9 +75,15 @@ class Client
         $hmac = new Hmac($this->token);
         $request = $hmac->signRequest($request);
 
-        $client = new FileGetContents($factory);
-        $response = $client->sendRequest($request);
-        return new Response($response);
+        $client = $this->networkClient ?: new FileGetContents($factory);
+        $networkResponse = $client->sendRequest($request);
+        $response = new Response($networkResponse);
+
+        if ($response->getStatusCode() >= 400) {
+            throw new ResponseException($response->getData()->message);
+        }
+
+        return $response;
     }
 
     /**