+++ src/Exception/ResponseException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Kuusamo\Api\Exception;
+
+use Exception;
+
+class ResponseException extends Exception
+{
+}