+++ test/ResponseTest.php
@@ -14,9 +14,11 @@ class ResponseTest extends TestCase
 
         $responseMock = $this->createMock('Nyholm\Psr7\Response');
         $responseMock->method('getBody')->willReturn($streamMock);
+        $responseMock->method('getStatusCode')->willReturn(200);
 
         $response = new Response($responseMock);
 
         $this->assertSame(['hello', 'world'], $response->getData());
+        $this->assertSame(200, $response->getStatusCode());
     }
 }