+++ changelog.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+2.1
+---
+* `Response` now has a `getStatusCode` method
+* Failed requests throw a `ResponseException`
+
 2.0.2
 -----
 * Stream can only be read once